refactor!: Stop exporting PlaybackState via mopidy.audio.PlaybackState
Use mopidy.types.PlaybackState instead.
