+++ src/mopidy/core/actor.py
@@ -13,7 +13,6 @@ from pykka.typing import ActorMemberMixin, proxy_method
 
 import mopidy
 from mopidy import audio, backend, mixer
-from mopidy.audio import PlaybackState
 from mopidy.core.history import HistoryController
 from mopidy.core.library import LibraryController
 from mopidy.core.listener import CoreListener
@@ -23,6 +22,7 @@ from mopidy.core.playlists import PlaylistsController
 from mopidy.core.tracklist import TracklistController
 from mopidy.internal import path, storage
 from mopidy.internal.models import CoreControllersState, StoredState
+from mopidy.types import PlaybackState
 
 if TYPE_CHECKING:
     from mopidy.config import Config
